detector/pyproject.toml
[project]
name = "anomaly-detection-opme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.114.0",
    "crewai-tools>=0.45.0",
    "docling>=2.32.0",
    "langchain>=0.3.25",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "reportlab>=4.4.1",
]
